Moved TagManager to page directory, fixed imports
